% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{aci}
\alias{aci}
\title{Initialize an Adaptive Conformal Inference algorithm.}
\usage{
aci(
  Y = NULL,
  predictions = NULL,
  training = FALSE,
  alpha = 0.95,
  method = "ACI",
  parameters = list()
)
}
\arguments{
\item{Y}{optional vector of observations}

\item{predictions}{optional vector of predictions}

\item{training}{optional boolean indicating if the supplied Y and prediction values should be treated as training data (not used to update ACI parameters, but used in e.g. calculating conformity scores)}

\item{alpha}{desired level o}

\item{method}{a string specificying the Adaptive Conformal Inference method to use.
The available methods are:
\describe{
   \item{'ACI'}{The original Adaptive Conformal Inference method proposed by \insertCite{gibbs2021aci}.
   Requires specification of a positive learning rate \strong{gamma}.}
   \item{'AgACI'}{Aggregated ACI method introduced by \insertCite{zaffran2022agaci}.
   Multiple ACI algorithms are executed for a grid of learning rates, and the resulting intervals are combined using the Bernstein Online Aggregation method
   for online aggregation of experts. Requires specification of a grid of learning rates \strong{gamma_grid}.}
}}

\item{parameters}{a list of parameters that depend on the chosen ACI method.
\describe{
   \item{gamma}{A positive number specifying the learning rate. For method "ACI".}
   \item{gamma_grid}{A grid of positive learning rates. For method "AgACI".}
   \item{interval_constructor}{
     Specifies how the prediction intervals are to be formed.
     \begin{itemize}
       \item "linear": the interval is formed as [prediction - theta, prediction + theta].
       \item "conformal": the interval is formed as [prediction - S, prediction + S], where S is the theta*100% quantile of the previously observed
       conformity scores.
       \item A function that takes a prediction and a parameter theta (that is learned online) and returns a candidate prediction interval.
     \end{itemize}
   }
   \item{conformity_score}{
     Specifies the conformity scores to use for constructing prediction intervals. Only applicable if interval_constructor = "conformal".
     Current options are:
     \begin{align}
       \item "absolute_error": S = |Y - prediction|
       \item A function taking an observation Y and prediction and returning a conformity score.
     \end{align}
   }
}}
}
\description{
Initialize an Adaptive Conformal Inference algorithm.
}
