% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{aci_metrics}
\alias{aci_metrics}
\title{Compute metrics for ACI prediction intervals.}
\usage{
aci_metrics(object, indices = NULL)
}
\arguments{
\item{object}{ACI object}

\item{indices}{indices of observations to include in computation of metrics}
}
\description{
Compute metrics for ACI prediction intervals.
}
