% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformity_scores.R
\name{conformity_score_absolute_error}
\alias{conformity_score_absolute_error}
\title{Absolute error conformity score: S_t := |Y_t - \hat{Y}_t|}
\usage{
conformity_score_absolute_error(Y, prediction)
}
\arguments{
\item{Y}{observed value}

\item{prediction}{predicted value}
}
\description{
Absolute error conformity score: S_t := |Y_t - \hat{Y}_t|
}
