% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.aci}
\alias{plot.aci}
\title{Plot an ACI object.}
\usage{
\method{plot}{aci}(x, index = NULL, legend = TRUE, predictions = TRUE, engine = "base", ...)
}
\arguments{
\item{x}{object of class "aci"}

\item{index}{indexes of the observation and intervals to plot (defaults to showing all observations and intervals).}

\item{legend}{logical indicating whether or not to include a legend}

\item{predictions}{whether or not to include the original predictions in plot}

\item{engine}{either "base" or "ggplot"}

\item{...}{additional arguments to base plot function}
}
\description{
Plot an ACI object.
}
