% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_agaci_weights}
\alias{plot_agaci_weights}
\title{Plot most recent weights assigned to each of the candidate ACI algorithms
by the AgACI method}
\usage{
plot_agaci_weights(x, type = "last", legend = "topright", ...)
}
\arguments{
\item{x}{object of class "aci"}

\item{type}{type of plot: "last" to show only most recent weights, "last" to show all weights}

\item{legend}{legend position}

\item{...}{additional arguments to plot}
}
\description{
Plot most recent weights assigned to each of the candidate ACI algorithms
by the AgACI method
}
