% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{predict.aci}
\alias{predict.aci}
\title{Compute a conformal prediction interval}
\usage{
\method{predict}{aci}(object, prediction)
}
\arguments{
\item{object}{object of class "aci"}

\item{prediction}{vector or matrix of predictions to use for forming the conformal prediction interval}
}
\description{
Compute a conformal prediction interval
}
