% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{print.aci}
\alias{print.aci}
\title{Print an ACI object}
\usage{
\method{print}{aci}(x, ...)
}
\arguments{
\item{x}{object of class "aci"}

\item{...}{additional arguments (currently ignored)}
}
\description{
Print an ACI object
}
