% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{summary.aci}
\alias{summary.aci}
\title{Summary of an ACI object}
\usage{
\method{summary}{aci}(object, ...)
}
\arguments{
\item{object}{object of class "aci"}

\item{...}{additional arguments (currently ignored)}
}
\description{
Summary of an ACI object
}
