% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.aci}
\alias{update.aci}
\title{Update an ACI object with a new observation}
\usage{
\method{update}{aci}(object, newY, newpredictions, training = FALSE)
}
\arguments{
\item{object}{object of class "aci"}

\item{newY}{new observation}

\item{newpredictions}{new predictions}
}
\description{
Update an ACI object with a new observation
}
