% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{aci}
\alias{aci}
\title{Initialize an Adaptive Conformal Inference algorithm.}
\usage{
aci(
  Y = NULL,
  predictions = NULL,
  X = NULL,
  training = FALSE,
  alpha = 0.95,
  method = "AgACI",
  parameters = list()
)
}
\arguments{
\item{Y}{optional vector of observations.}

\item{predictions}{optional vector or matrix of predictions.}

\item{X}{optional matrix of covariates}

\item{training}{optional boolean indicating if the supplied Y and prediction values should be treated as training data (not used to update ACI parameters, but used in e.g. calculating conformity scores).}

\item{alpha}{desired empirical coverage. That is, the method will target alpha * 100\% prediction intervals.}

\item{method}{a string specifying the Adaptive Conformal Inference method to use.
The available methods are:
\describe{
   \item{'ACI'}{Adaptive Conformal Inference \insertCite{gibbs2021aci}{AdaptiveConformal}.}
   \item{'SCP'}{Split Conformal Prediction.}
   \item{'AgACI'}{Aggregated ACI \insertCite{zaffran2022agaci}{AdaptiveConformal}.
   Multiple ACI algorithms are executed for a grid of learning rates, and the resulting intervals are combined using Bernstein Online Aggregation.}
   \item{'FACI'}{Fully Adaptive Conformal Inference \insertCite{gibbs2022faci}{AdaptiveConformal}.}
   \item{'SF-OGD'}{Scale-Free Online Gradient Descent \insertCite{orabona2018scalefree}{AdaptiveConformal}.}
   \item{'SAOCP'}{Strongly Adaptive Online Conformal Prediction \insertCite{bhatnagar2023improved}{AdaptiveConformal}.}
}}

\item{parameters}{a list of parameters that depend on the chosen ACI method.
\describe{
   \item{gamma}{A positive number specifying the learning rate. For method "ACI".}
   \item{gamma_grid}{A grid of positive learning rates. For method "AgACI" and "FACI.}
   \item{theta0}{Initial value for theta parameter.}
   \item{interval_constructor}{
     Specifies how the prediction intervals are to be formed. Only for "ACI" and "AgACI". For "FACI", it is always set to "conformal".
     \itemize{
       \item "conformal": the interval is formed as [prediction - S, prediction + S], where S is the theta*100\% quantile of the previously observed.
       This is the default for "ACI" and "AgACI".
       \item "linear": the interval is formed as [prediction - theta, prediction + theta].
       conformity scores.
       \item A function (see details)
     }
   }
   \item{conformity_score}{
     Specifies the conformity scores to use for constructing prediction intervals. Only applicable if interval_constructor = "conformal".
     Current options are:
     \itemize{
       \item "absolute_error": S_t = |Y_t - prediction_t|. This is the default.
       \item A function (see details)
     }
   }
   \item{symmetric}{
     A boolean indicating whether the intervals should be symmetric (the default) or asymmetric.
   }
   \item{conditional}{
     A boolean indicating whether the intervals should be generated conditional on covariates.
   }
}}
}
\value{
An object with class "aci". Important parameters include:
\describe{
  \item{intervals}{A matrix that contains the prediction intervals generated by the method.}
  \item{covered}{A logical vector indicating if each observation was within (TRUE) or outside (FALSE) the prediction interval.}
  \item{training}{A logical vector indicating if each observation and prediction was used solely as context for the method (TRUE)
  or if they were used to update the underlying ACI method (FALSE)}.
}
}
\description{
Initialize an Adaptive Conformal Inference algorithm.
}
\details{
The prediction intervals are formed by an \emph{interval constructor} function.
  A custom interval constructor can be supplied as a function of the form
  \code{function(prediction, theta, object)} where \code{prediction} is the predicted value of the outcome,
  \code{theta} is a parameter that controls the size of the interval, and \code{object} is an ACI object.

  The "conformal" interval constructor forms intervals using quantiles of the previous \emph{conformity scores} of
  the predicted outcomes. The default conformity score is the absolute residual: S_t = |Y_t - prediction_t|, where S_t is the
  conformity score, Y_t is the observed value, and prediction_t is the predicted value.
  A custom conformity score can be supplied as a function of the form
  \code{function(Y, prediction)}.
}
\examples{
# Generate a simple time series of observations
N <- 1e3
Y <- rnorm(N)
predictions <- rep(0, N) # predict 0 at very timestep

# Apply ACI/Rolling Risk Control algorithm with learning rate gamma = 0.01
# with goal of 90\\% prediction intervals
result <- aci(Y, predictions, method = "ACI", alpha = 0.9,
              parameters = list(gamma = 0.01))
summary(result)

# Supply a custom conformity score function
conformity_score_squared_error <- function(Y, prediction) (Y - prediction)^2
result <- aci(Y, predictions, method = "ACI", alpha = 0.9,
              parameters = list(gamma = 0.01, conformity_score = conformity_score_squared_error))
summary(result)

}
\references{
\insertAllCited{}
}
