% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{predict.aci}
\alias{predict.aci}
\title{Compute a conformal prediction interval}
\usage{
\method{predict}{aci}(object, prediction = 0, X = NULL, ...)
}
\arguments{
\item{object}{object of class "aci"}

\item{prediction}{vector or matrix of predictions to use for forming the conformal prediction interval}

\item{X}{matrix of covariates}

\item{...}{additional arguments (currently ignored)}
}
\description{
Compute a conformal prediction interval
}
